Integrate Issue Tracker with the new bug tracker. Refs #699
- moved IssueTrackerType into a base module to simplify
  types.py
- renamed IssueTrackerType.from_name() as types.from_name()
- IssueTrackerType.report_issue_from_testexecution() now has
  a second parameter ``user``
- added tcms.bugs.views.New.create_bug() helper method which
  also tries to automatically find assignee
- added kiwitcms.py which doesn't go via the RPC layer but
  instead works internally
- API method TestExecution.add_link() special cases this new IT
  and always updates the references between Bug & TestExecution
  models
- when clicking 'Report bug' from inside TestExecution the new
  bug is reported automatically and new browser window opens to
  display the bug information
- OpenGraph support for bugs/get.html template
