+++ docs/source/modules/tcms.issuetracker.base.rst
@@ -1,0 +1,7 @@
+tcms.issuetracker.base module
+=============================
+
+.. automodule:: tcms.issuetracker.base
+   :members:
+   :undoc-members:
+   :show-inheritance:
