--- docs/source/modules/tcms.issuetracker.base_integration.rst
@@ -1,7 +1,0 @@
-tcms.issuetracker.base\_integration module
-==========================================
-
-.. automodule:: tcms.issuetracker.base_integration
-   :members:
-   :undoc-members:
-   :show-inheritance:
