+++ docs/source/modules/tcms.issuetracker.kiwitcms.rst
@@ -1,0 +1,7 @@
+tcms.issuetracker.kiwitcms module
+=================================
+
+.. automodule:: tcms.issuetracker.kiwitcms
+   :members:
+   :undoc-members:
+   :show-inheritance:
