+++ tcms/xmlrpc/api/utils.py
@@ -2,7 +2,7 @@
 
 # Licensed under the GPL 2.0: https://www.gnu.org/licenses/old-licenses/gpl-2.0.html
 
-from tcms.issuetracker.types import IssueTrackerType
+from tcms.issuetracker.types import from_name
 from tcms.testcases.models import BugSystem
 
 
@@ -13,7 +13,7 @@ def tracker_from_url(url):
         URLs to pre-existing defects to this method.
     """
     for bug_system in BugSystem.objects.all():
-        if url.startswith(bug_system.base_url):
-            return IssueTrackerType.from_name(bug_system.tracker_type)(bug_system)
+        if bug_system.base_url and url.startswith(bug_system.base_url):
+            return from_name(bug_system.tracker_type)(bug_system)
 
     return None
